server/pyproject.toml
[project]
name = "highstake-server"
version = "0.2.0"
description = "HighStake AI-powered boardroom simulator backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-socketio>=5.11.0",
    "sqlalchemy[asyncio]>=2.0.30",
    "aiosqlite>=0.20.0",
    "alembic>=1.13.0",
    "pydantic-settings>=2.2.0",
    "python-multipart>=0.0.9",
    "boto3>=1.34.0",
    "httpx>=0.27.0",
    "anthropic>=0.40.0",
    "deepgram-sdk>=3.4.0",
    "python-pptx>=0.6.23",
    "PyMuPDF>=1.24.0",
    "Pillow>=10.3.0",
    "aiofiles>=24.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
